refactor(X-09): extract TripRepository as abstract interface
Split the concrete class into an abstract TripRepository contract and
FirestoreTripRepository implementation. Provider now exposes the abstract
type so that tests and dev stubs can override without touching feature code.
